+++ Sources/SuperwallKit/Permissions/Handlers/Contacts/PermissionsHandler+Contacts.swift
@@ -15,13 +15,10 @@ extension PermissionHandler {
     return raw.toContactsPermissionStatus
   }
 
+  @MainActor
   func requestContactsPermission() async -> PermissionStatus {
     guard hasPlistKey(PlistKey.contacts) else {
-      Logger.debug(
-        logLevel: .error,
-        scope: .paywallViewController,
-        message: "Missing \(PlistKey.contacts) in Info.plist. Cannot request contacts permission."
-      )
+      await showMissingPlistKeyAlert(for: PlistKey.contacts, permissionName: "Contacts")
       return .unsupported
     }